// comonl-comonlwcl/src/app/modules/comonlapi/model/delega.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApiError } from './apiError';
import { Comune } from './comune';
import { Personalizzazione } from './personalizzazione';
import { StatiEsteri } from './statiEsteri';
import { StatoDelega } from './statoDelega';
import { TipoDelega } from './tipoDelega';


export interface Delega { 
    apiWarnings?: Array<ApiError>;
    id?: number;
    annoProtocollo?: number;
    capDelegante?: string;
    capDelegato?: string;
    carica?: string;
    cfDelegante?: string;
    cfDelegato?: string;
    cfImpresa?: string;
    cognomeDelegante?: string;
    cognomeDelegato?: string;
    comuneDelegante?: string;
    comuneDelegato?: string;
    denominazioneImpresa?: string;
    dtInsert?: Date;
    dtProtocollo?: Date;
    dtRevoca?: Date;
    dtUltMod?: Date;
    dtVerifica?: Date;
    faxDelegato?: string;
    idUserInsert?: string;
    idUserUltMod?: string;
    indirizzoDelegante?: string;
    indirizzoDelegato?: string;
    mailDelegante?: string;
    mailDelegato?: string;
    numeroProtocollo?: number;
    nomeDelegante?: string;
    nomeDelegato?: string;
    pivaDelegato?: string;
    pivaImpresa?: string;
    pvDelegante?: string;
    pvDelegato?: string;
    telefonoDelegante?: string;
    telefonoDelegato?: string;
    verifica?: string;
    comDelegato?: Comune;
    comDelegante?: Comune;
    statiEsteriDelegato?: StatiEsteri;
    statiEsteriDelegante?: StatiEsteri;
    personalizzazione?: Personalizzazione;
    statoDelega?: StatoDelega;
    tipoDelega?: TipoDelega;
    cente?: string;
}