// comonl-comonlwcl/src/app/modules/comonlapi/model/statiEsteri.ts
/**
 * ComonlWeb
 * API per ComonlWeb
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApiError } from './apiError';
import { DatiAzienda } from './datiAzienda';
import { LavoratoreSilp } from './lavoratoreSilp';
import { Sede } from './sede';


export interface StatiEsteri { 
    apiWarnings?: Array<ApiError>;
    id?: number;
    codNazioneMin?: string;
    dsStatiEsteri?: string;
    dtFine?: Date;
    dtInizio?: Date;
    dtTmst?: Date;
    flgUe?: string;
    siglaNazione?: string;
    datiAziendas?: Array<DatiAzienda>;
    lavoratoriSilps?: Array<LavoratoreSilp>;
    sedeLegales?: Array<Sede>;
}